Fixes #117. Moves transforms to own files
