+++ src/plugin/figmaTransforms/lineHeight.ts
@@ -1,0 +1,34 @@
+import {convertTypographyNumberToFigma} from './generic';
+
+export function convertLineHeightToFigma(inputValue) {
+    let lineHeight;
+    const value = inputValue.toString();
+    const numbers = /^\d+(\.\d+)?$/;
+    if (value.match(numbers) || value.endsWith('px')) {
+        lineHeight = {
+            unit: 'PIXELS',
+            value: convertTypographyNumberToFigma(value),
+        };
+    } else if (value.trim().slice(-1) === '%' && value.trim().slice(0, -1).match(numbers)) {
+        lineHeight = {
+            unit: 'PERCENT',
+            value: Number(value.slice(0, -1)),
+        };
+    } else {
+        lineHeight = {
+            unit: 'AUTO',
+        };
+    }
+    return lineHeight;
+}
+
+export function convertFigmaToLineHeight(inputValue) {
+    const {unit, value} = inputValue;
+    if (unit === 'PIXELS') {
+        return +value.toFixed(2);
+    }
+    if (unit === 'PERCENT') {
+        return `${+value.toFixed(2)}%`;
+    }
+    return 'AUTO';
+}
