+++ src/plugin/helpers.test.ts
@@ -1,13 +1,4 @@
-import {
-    isObject,
-    mergeDeep,
-    convertFigmaToLineHeight,
-    hslaToRgba,
-    convertToFigmaColor,
-    hexToRgb,
-    convertStringToFigmaGradient,
-    convertFigmaGradientToString,
-} from './helpers';
+import {isObject, mergeDeep, transformValue} from './helpers';
 
 describe('isObject', () => {
     it('correctly asserts an object', () => {
@@ -32,272 +23,77 @@ describe('mergeDeep', () => {
     });
 });
 
-describe('convertFigmaToLineHeight', () => {
-    it('converts a figma line height to a readable input value', () => {
-        const lineHeightPx = convertFigmaToLineHeight({unit: 'PIXELS', value: 13});
-        expect(lineHeightPx).toBe(13);
-        const lineHeightPerc = convertFigmaToLineHeight({unit: 'PERCENT', value: 13});
-        expect(lineHeightPerc).toBe('13%');
-        const lineHeightAuto = convertFigmaToLineHeight({unit: 'AUTO'});
-        expect(lineHeightAuto).toBe('AUTO');
-    });
-});
-
-describe('hslaToRgba', () => {
-    it('converts hsla to rgba', () => {
-        const hsl = hslaToRgba([210, 50, 50]);
-        expect(hsl).toEqual([64, 128, 191, 1]);
-        const hsla = hslaToRgba([210, 50, 50, 0.5]);
-        expect(hsla).toEqual([64, 128, 191, 0.5]);
-    });
-});
-describe('hexToRgb', () => {
-    it('converts hex to rgb', () => {
-        const color = '#ff0000';
-
-        expect(hexToRgb(color)).toEqual({r: 255, g: 0, b: 0});
-    });
-
-    it('returns null if no match', () => {
-        const color = 'rgb(0000)';
-
-        expect(hexToRgb(color)).toBe(null);
-    });
-});
-
-describe('convertToFigmaColor', () => {
-    it('convert any color to a figma readable color', () => {
-        const rgb = 'rgb(255,255,255)';
-        expect(convertToFigmaColor(rgb)).toEqual({
-            color: {
-                r: 1,
-                g: 1,
-                b: 1,
+describe('transformValue', () => {
+    const tokens = [
+        {
+            input: '12px',
+            type: 'spacing',
+            output: 12,
+        },
+        {
+            input: '12px',
+            type: 'lineHeights',
+            output: {
+                unit: 'PIXELS',
+                value: 12,
             },
-            opacity: 1,
-        });
-        const rgba = 'rgba(51,51,51,0.5)';
-        expect(convertToFigmaColor(rgba)).toEqual({
-            color: {
-                r: 0.2,
-                g: 0.2,
-                b: 0.2,
+        },
+        {
+            input: '160%',
+            type: 'lineHeights',
+            output: {
+                unit: 'PERCENT',
+                value: 160,
             },
-            opacity: 0.5,
-        });
-        const hex = '#ff0000';
-        expect(convertToFigmaColor(hex)).toEqual({
-            color: {
-                r: 1,
-                g: 0,
-                b: 0,
+        },
+        {
+            input: '24px',
+            type: 'letterSpacing',
+            output: {
+                unit: 'PIXELS',
+                value: 24,
             },
-            opacity: 1,
-        });
-        const hexa = '#ff0000cc';
-        expect(convertToFigmaColor(hexa)).toEqual({
-            color: {
-                r: 1,
-                g: 0,
-                b: 0,
+        },
+        {
+            input: '120%',
+            type: 'letterSpacing',
+            output: {
+                unit: 'PERCENT',
+                value: 120,
             },
-            opacity: 0.8,
+        },
+        {
+            input: '120%',
+            type: 'lineHeights',
+            output: {
+                unit: 'PERCENT',
+                value: 120,
+            },
+        },
+        {
+            input: '2rem',
+            type: 'sizing',
+            output: 32,
+        },
+        {
+            input: '0.24',
+            type: 'sizing',
+            output: 0.24,
+        },
+        {
+            input: '50%',
+            type: 'opacity',
+            output: 0.5,
+        },
+        {
+            input: '0.6',
+            type: 'opacity',
+            output: 0.6,
+        },
+    ];
+    it('transforms non-conform values into their required formats', () => {
+        tokens.forEach((token) => {
+            expect(transformValue(token.input, token.type)).toEqual(token.output);
         });
     });
-});
-
-describe('convertStringtoFigmaGradient', () => {
-    const test1 = {
-        input: 'linear-gradient(45deg, #ffffff 0%, #000000 100%)',
-        output: {
-            gradientStops: [
-                {
-                    color: {
-                        a: 1,
-                        b: 1,
-                        g: 1,
-                        r: 1,
-                    },
-                    position: 0,
-                },
-                {
-                    color: {
-                        a: 1,
-                        b: 0,
-                        g: 0,
-                        r: 0,
-                    },
-                    position: 1,
-                },
-            ],
-            gradientTransform: [
-                [0.71, 0.71, 0],
-                [-0.71, 0.71, 0],
-            ],
-        },
-    };
-
-    const test2 = {
-        input: 'linear-gradient(45deg, #ffffff 0%, rgba(255,0,0,0.5) 50%, #000000 100%)',
-        output: {
-            gradientStops: [
-                {
-                    color: {
-                        a: 1,
-                        b: 1,
-                        g: 1,
-                        r: 1,
-                    },
-                    position: 0,
-                },
-                {
-                    color: {
-                        a: 0.5,
-                        b: 0,
-                        g: 0,
-                        r: 1,
-                    },
-                    position: 0.5,
-                },
-                {
-                    color: {
-                        a: 1,
-                        b: 0,
-                        g: 0,
-                        r: 0,
-                    },
-                    position: 1,
-                },
-            ],
-            gradientTransform: [
-                [0.71, 0.71, 0],
-                [-0.71, 0.71, 0],
-            ],
-        },
-    };
-
-    const test3 = {
-        input: 'linear-gradient(45deg, #ffffff 0%, rgba(255,0,0,0.5) 50%, #000000 100%)',
-        output: {
-            gradientStops: [
-                {
-                    color: {
-                        a: 1,
-                        b: 1,
-                        g: 1,
-                        r: 1,
-                    },
-                    position: 0,
-                },
-                {
-                    color: {
-                        a: 0.5,
-                        b: 0,
-                        g: 0,
-                        r: 1,
-                    },
-                    position: 0.5,
-                },
-                {
-                    color: {
-                        a: 1,
-                        b: 0,
-                        g: 0,
-                        r: 0,
-                    },
-                    position: 1,
-                },
-            ],
-            gradientTransform: [
-                [0.71, 0.71, 0],
-                [-0.71, 0.71, 0],
-            ],
-        },
-    };
-
-    expect(convertStringToFigmaGradient(test1.input)).toEqual(test1.output);
-    expect(convertStringToFigmaGradient(test2.input)).toEqual(test2.output);
-    expect(convertStringToFigmaGradient(test3.input)).toEqual(test3.output);
-});
-
-describe('convertFigmaGradientToString', () => {
-    const test1: {
-        input: GradientPaint;
-        output: string;
-    } = {
-        input: {
-            type: 'GRADIENT_LINEAR',
-            gradientStops: [
-                {
-                    color: {
-                        a: 1,
-                        b: 1,
-                        g: 1,
-                        r: 1,
-                    },
-                    position: 0,
-                },
-                {
-                    color: {
-                        a: 1,
-                        b: 0,
-                        g: 0,
-                        r: 0,
-                    },
-                    position: 1,
-                },
-            ],
-            gradientTransform: [
-                [0.71, 0.71, 0],
-                [0.71, 0.71, 0],
-            ],
-        },
-        output: 'linear-gradient(45deg, #ffffff 0%, #000000 100%)',
-    };
-
-    const test2: {
-        input: GradientPaint;
-        output: string;
-    } = {
-        input: {
-            type: 'GRADIENT_LINEAR',
-            gradientStops: [
-                {
-                    color: {
-                        a: 1,
-                        b: 1,
-                        g: 1,
-                        r: 1,
-                    },
-                    position: 0,
-                },
-                {
-                    color: {
-                        a: 1,
-                        b: 0,
-                        g: 0,
-                        r: 1,
-                    },
-                    position: 0.535,
-                },
-                {
-                    color: {
-                        a: 1,
-                        b: 0,
-                        g: 0,
-                        r: 0,
-                    },
-                    position: 1,
-                },
-            ],
-            gradientTransform: [
-                [0.71, 0.71, 0],
-                [0.71, 0.71, 0],
-            ],
-        },
-        output: 'linear-gradient(45deg, #ffffff 0%, #ff0000 53.5%, #000000 100%)',
-    };
-
-    expect(convertFigmaGradientToString(test1.input)).toEqual(test1.output);
-    expect(convertFigmaGradientToString(test2.input)).toEqual(test2.output);
 });
